organized files and added app.get()'s
