--- index.js
@@ -1,16 +1,0 @@
-const express = require('express');
-const app = express();
-const path = require('path');
-const server = require('http').createServer(app);
-const io = require('socket.io')(server);
-const port = process.env.PORT || 3000;
-
-server.listen(port, () => {
-  console.log('Server listening at port %d', port);
-});
-
-app.use(express.static(path.join(__dirname, 'public')));
-
-let numUsers = 0;
-let clientList = [];
-
