+++ package.json
@@ -2,10 +2,10 @@
   "name": "cst336-tp1",
   "version": "1.0.0",
   "description": "This is a small game made for CST 336 at CSU Monterey",
-  "main": "index.js",
+  "main": "app.js",
   "scripts": {
-    "start": "node index.js",
-    "dev": "nodemon index.js"
+    "start": "node app.js",
+    "dev": "nodemon app.js"
   },
   "repository": {
     "type": "git",
@@ -18,6 +18,8 @@
   },
   "homepage": "https://github.com/dsullivan89/cst336-tp1#readme",
   "dependencies": {
+    "bluebird": "^3.7.2",
+    "ejs": "^3.1.5",
     "express": "^4.17.1",
     "nodemon": "^2.0.6",
     "socket.io": "^3.0.1"
