+++ public/css/styles.css
@@ -1,3 +1,18 @@
+/* Centering our main content horizontally and vertically */
+html,
+body {
+  height: 100%;
+  width: 100%;
+}
+
+.container {
+  align-items: center;
+  display: flex;
+  justify-content: center;
+  height: 100%;
+  width: 100%;
+}
+
 #dialog-overlay {
 	background: #000;
 	bottom: 0;
@@ -13,6 +28,40 @@
 	width: 100%;
 	z-index: 99998;
 }
+
+.lounge {
+	width: 640px;
+	height: 480px;
+}
+
+body {
+	background-color: black;
+}
+
+h1, h2, h3 {
+	color: chocolate;
+}
+
+.border-g {
+	border-color: green;
+}
+
+.border-w {
+	border-color: white;
+}
+
+.border-y {
+	border-color: yellow;
+}
+
+.border-thin {
+	border: 2px solid;
+}
+
+.border-thick {
+	border: 4px solid;
+}
+
 
 .dialog-popup {
 	display: none;
