--- public/login.html
@@ -1,29 +1,0 @@
-<!DOCTYPE html>
-<html>
-	<head>
-		<title>Login</title>
-	</head>
-	<body>
-		<!-- register / login links (located near the top right -->
-		<div id="topLinks">
-			<a href="register.html">Register</a>
-			<a href="login.html">Login</a>
-		</div>
-		<!-- Main Header-->
-		<h1>Welcome back.</h1>
-		<p>If you are a returning user please login. If you are new here we invite you to
-			<a href="register.html">register</a> for a free account. </p>
-		<br><br>
-		<h2>Login</h2>
-		<div id="login-dialog" class="dialog-popup">
-			<form method="post" action="#">
-				<input type="text" id="accountName" placeholder="account name">
-				<input type="password" id="password" placeholder="password">
-				<button class="submit" type="button">Login</button>
-			</form>
-		</div>
-	</body>
-	<footer>
-		<h6>Monterey Bay CST 336 - Team Project 1</h6>
-	</footer>
-</html>
