--- public/register.html
@@ -1,30 +1,0 @@
-<!DOCTYPE html>
-<html>
-	<head>
-		<title>Login</title>
-	</head>
-	<body>
-		<!-- register / login links (located near the top right -->
-		<div id="topLinks">
-			<a href="register.html">Register</a>
-			<a href="login.html">Login</a>
-		</div>
-		<!-- Main Header-->
-		<h1>Welcome Hero</h1>
-		<p>If you already have an account you wish to use then you may
-			<a href="login.html">login</a> here. </p>
-		<br><br>
-		<h2>Register</h2>
-		<form method="post" action="#">
-			<input type="text" id="accountName" placeholder="account name"> <br>
-			<input type="password" id="password" placeholder="password"> <br>
-			<input type="password" id="repeatPassword" placeholder="reenter password"> <br>
-			<input type="text" id="displayName" placeholder="Display name"> 
-			<button class="submit" type="button">Register</button>
-		</form>
-		<p>players will know you by your display name. Never give out your account name or password.</p>
-	</body>
-	<footer>
-		<h6>Monterey Bay CST 336 - Team Project 1</h6>
-	</footer>
-</html>
