+++ views/authenticated/lounge.html
@@ -1,0 +1,24 @@
+<!DOCTYPE html>
+<html lang="en">
+	<head>
+		<title>CST-336 Midterm</title>
+		<meta charset="UTF-8">
+		<link rel="stylesheet" 
+		 href="https://stackpath.bootstrapcdn.com/bootstrap/4.5.2/css/bootstrap.min.css" 
+		 integrity="sha384-JcKb8q3iqJ61gNV9KGb8thSsNjpSL0n8PARn9HuZOnIxN0hoP+VmmDGMN5t9UJ0Z" 
+		 crossorigin="anonymous">
+	 	<link rel="stylesheet" href="../../css/styles.css" type="text/css">
+	 	<script src="https://ajax.googleapis.com/ajax/libs/jquery/3.5.1/jquery.min.js"></script>
+	 	<script src="../../js/lounge.js" type="text/javascript"></script>
+	</head>
+	<body>
+		<div class="container">
+			<div class="lounge border-thick border-g">
+				<h1>You've entered the lounge!</h1>
+				<a href="/lobby">Create</a>
+			</div>
+		</div>
+  </body>
+  <footer>
+
+  </footer>
